move dpm tests into new file & fix unit tests
